packages/qphase/requirements.txt
typer>=0.9
rich>=13.0
ruamel.yaml>=0.17
PyYAML>=6.0
pydantic>=2.0
